pyproject.toml
[tool.poetry]
name = "pobr-ffmpeg"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
ffmpeg-python = "^0.2.0"
numpy = "^1.22.4"
opencv-python = "^4.5.5"
black = {extras = ["jupyter"], version = "^22.3.0"}
tqdm = "^4.66.1"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
